Undo a previous signedness fix This had the effect of causing a different comparison issue. Instead of trying to untangle this function that could probably be greatly simplified, we'll just cast size() and hope for the best.

+++ soh/soh/Enhancements/debugger/valueViewer.cpp
@@ -115,7 +115,7 @@ void ValueViewerWindow::DrawElement() {
         std::string(valueTable[selectedElement].name) + " (" + std::string(valueTable[selectedElement].path) + ")"
     );
     if (ImGui::BeginCombo("##valueViewerElement", selectedElementText.c_str())) {
-        for (size_t i = 0; i < valueTable.size(); i++) {
+        for (int i = 0; i < static_cast<int>(valueTable.size()); i++) {
             if (valueTable[i].isActive) continue;
             bool isSelected = (selectedElement == i);
             std::string elementText = (
